made read code into library
