+++ GOES/GOES-17/GLM/Makefile.am
@@ -9,11 +9,17 @@ LOG_DRIVER = $(SHELL) $(top_srcdir)/test-driver-verbose
 
 TEST_EXTENSIONS = .sh
 
+# This is a libtool library.
+lib_LTLIBRARIES = libgoesglm.la
+libgoesglm_la_LDFLAGS = -version-info 0:0:0
+libgoesglm_la_SOURCES = glm_read.c glm_data.h
+
 # Build our test program.
-check_PROGRAMS = glm_read
+check_PROGRAMS = tst_glm_read
 
 # Source code for the test.
-glm_read_SOURCES = glm_read.c glm_data.h un_test.h tst_utils.c
+tst_glm_read_SOURCES = tst_glm_read.c un_test.h tst_utils.c
+tst_glm_read_LDADD = libgoesglm.la
 
 # Run our test program.
 TESTS = tst_glm.sh
